refactor: move secrets mapping to settingspage
